requirements.txt
streamlit
langchain
python-dotenv
langchain_google_genai
